requirements.txt
apscheduler
litellm
gitpython
PyGithub
watchfiles
rich
plyer
pyobjus
sqlite-utils
tiktoken
python-dotenv
fastapi
uvicorn
pyngrok
